Bug 6090 - Exception thrown from datastore as soon as a radio head is connected after fresh start of the controller
This issue is fixed by using a freshStart variable to keep track of whether the DeviceConnected event is trigger
for the first time; if so, do nothing and if not, delete all existing objects under the radio head node

Developer's Certificate of Origin 1.1

        By making a contribution to this project, I certify that:

        (a) The contribution was created in whole or in part by me and I
            have the right to submit it under the open source license
            indicated in the file; or

        (b) The contribution is based upon previous work that, to the best
            of my knowledge, is covered under an appropriate open source
            license and I have the right under that license to submit that
            work with modifications, whether created in whole or in part
            by me, under the same open source license (unless I am
            permitted to submit under a different license), as indicated
            in the file; or

        (c) The contribution was provided directly to me by some other
            person who certified (a), (b) or (c) and I have not modified
            it.

        (d) I understand and agree that this project and the contribution
            are public and that a record of the contribution (including all
            personal information I submit with it, including my sign-off) is
            maintained indefinitely and may be redistributed consistent with
            this project or the open source license(s) involved.

+++ applications/ocp-service/src/main/java/org/opendaylight/ocpplugin/applications/ocpservice/OcpService.java
@@ -82,6 +82,7 @@ public class OcpService implements OcpServiceService, DataChangeListener, SalDev
     private final SalDeviceMgmtService ocpDeviceMgmtService;
     private final SalObjectLifecycleService ocpObjectLifecycleService;    
     private static final InstanceIdentifier<Nodes> NODES_IDENTIFIER = InstanceIdentifier.create(Nodes.class);
+    private boolean freshStart = true;
     
     public OcpService(final RpcConsumerRegistry rpcRegistry) {
         executor = Executors.newFixedThreadPool(1);
@@ -481,9 +482,12 @@ public class OcpService implements OcpServiceService, DataChangeListener, SalDev
                 LOG.debug("Get deviceConnected notification with notification {}", notification);                  
                 
                 try {
-                    resourceModelBroker.deleteObj("ALL");
+                    if (!freshStart) {
+                        resourceModelBroker.deleteObj("ALL");
+                    }
+                    freshStart = false;
                 } catch (Exception exc) {
-                    LOG.error("Failed to delete node:({}, {})", notification.getNodeId(), exc); 
+                        LOG.info("Failed to delete node:({}, {})", notification.getNodeId(), exc);
                 }
                 try {
                     setInputBuilder.setNode(new NodeRef(NODES_IDENTIFIER.child(Node.class, new NodeKey(new NodeId(notification.getNodeId())))));
